Fix the slug for school events to include school id
This allows school events to be loaded in the single event view.
