+++ app/services/user-events.js
@@ -2,11 +2,11 @@ import Ember from 'ember';
 import config from 'ilios/config/environment';
 import EventMixin from 'ilios/mixins/events';
 import ajax from 'ic-ajax';
+import moment from 'moment';
 
 export default Ember.Service.extend(EventMixin, {
   store: Ember.inject.service(),
   currentUser: Ember.inject.service(),
-  slugType: 'U',
   getEvents(from, to){
     var deferred = Ember.RSVP.defer();
     this.get('currentUser.model').then(user => {
@@ -29,5 +29,40 @@ export default Ember.Service.extend(EventMixin, {
     });
 
     return deferred.promise;
-  }
+  },
+  getEventForSlug(slug){
+    let from = moment(slug.substring(1, 9), 'YYYYMMDD').hour(0);
+    let to = from.clone().hour(24);
+    let type = slug.substring(9, 10);
+    let id = parseInt(slug.substring(10));
+    
+    return new Ember.RSVP.Promise(resolved => {
+      this.getEvents(from.unix(), to.unix()).then(events => {
+        let event = events.find( event => {
+          if(type === 'O'){
+            return event.offering === id;
+          }
+          if(type === 'I'){
+            return event.ilmSession === id;
+          }
+        });
+        
+        resolved(event);
+      });
+    });
+    
+   
+  },
+  getSlugForEvent(event){
+    let slug = 'U';
+    slug += moment(event.startDate).format('YYYYMMDD');
+    if(event.offering){
+      slug += 'O' + event.offering;
+    }
+    if(event.ilmSession){
+      slug += 'I' + event.ilmSession;
+    }
+
+    return slug;
+  },
 });
